make the manager commit use common pkg
